refact: folder migrate
