+++ src/constants/accordionData.tsx
@@ -13,7 +13,7 @@ import GallerySelection from '../components/form/GallerySelection/GallerySelecti
 import LocationSelection from '../components/form/LocationSelection/LocationSelection';
 import AccountSelection from '../components/form/AccountSelection/AccountSelection';
 import ContactSelection from '../components/form/ContactSelection/ContactSelection';
-import GalleryForm from '../components/form/GalleryForm';
+import GallerySelecetion from '../components/form/GallerySelection/GallerySelection';
 
 export const accordionData: AccordionItemData[] = [
   {
@@ -109,6 +109,6 @@ export const accordionData: AccordionItemData[] = [
   {
     id: 19,
     title: '갤러리',
-    content: <GalleryForm />
+    content: <GallerySelecetion />
   },
 ];
